chore(sandcastle): inline workflow.ts into main.ts and harness.ts
workflow.ts existed to host config that used to be duplicated across
main.ts and prompt files, and to host the AgentSpec discriminated
union. With pipeline.ts gone there's only one consumer (main.ts) for
the constants, and the AgentSpec types are only consumed by harness.ts
(the dispatch site). The "single source of truth" justification turns
into "extra import hop" once the duplication is gone.

- Move LABEL, MAX_ITERATIONS, COPY_TO_WORKTREE, INSTALL_AND_VERIFY,
  AGENTS into main.ts at the top, grouped by purpose. The `satisfies
  Record<string, AgentSpec>` constraint still gives compile-time shape
  enforcement.
- Move the AgentSpec / ClaudeCodeSpec / CodexSpec / OpenCodeSpec /
  PiSpec types into harness.ts, next to the agent() switch that
  consumes them.
- Delete workflow.ts and workflow.test.ts. The runtime tests checked
  trivial things (LABEL non-empty, MAX_ITERATIONS positive integer)
  or what `satisfies` already enforces (every agent has a known
  provider, non-empty model). The promptPath-resolves check is
  redundant — sandcastle would fail loudly at run time on a missing
  file anyway.
