--- .sandcastle/workflow.ts
@@ -1,109 +1,0 @@
-// Single source of truth for sandcastle harness configuration.
-//
-// Values that historically lived inline in main.ts and were duplicated into
-// prompt files. Pulling them here means changing model/effort/labels happens
-// in one place; prompts receive the values via promptArgs.
-
-// ---------- Tracker conventions ----------
-
-// GitHub label used to opt issues into the Sandcastle workflow.
-export const LABEL = "Sandcastle";
-
-// ---------- Orchestrator limits ----------
-
-// One issue per iteration, sequentially. Parallelism was removed because the
-// only thing protecting it (the LLM-based subset selector) was an expensive
-// guess at file conflicts, and merge conflicts on overlapping PRs are a
-// normal git outcome — not worth a per-iteration LLM round-trip to avoid.
-export const MAX_ITERATIONS = 10;
-
-// ---------- Sandbox setup ----------
-
-// Skip host->worktree copy: this monorepo's node_modules is ~3.5GB and
-// blows past sandcastle's hard-coded 60s copy timeout. The implementer
-// sandbox runs `bun install` inside the container instead.
-// Env files are gitignored, so copy them in explicitly.
-export const COPY_TO_WORKTREE: readonly string[] = ["apps/desktop/.env", "apps/web/.env.local"];
-
-// Verify the install actually produced a usable workspace. `bun install` exits
-// 0 even when individual extractions are mangled, so we follow with `turbo
-// typecheck`, which resolves and loads imports across every workspace and
-// fails loudly if a package's main entry is missing.
-export const INSTALL_AND_VERIFY = "bun install && bun run typecheck";
-
-// ---------- Agent specs ----------
-
-// Provider-tagged spec. Each agent declares which sandcastle agent provider
-// it uses (claudeCode | codex | opencode | pi); main.ts dispatches on the
-// `provider` discriminator. Adding a new backend (e.g. Gemini, Ollama) means
-// adding a variant here and a case in main.ts's `agent()` helper, with no
-// other downstream changes.
-//
-// Per-provider `effort` types are sourced from sandcastle's own option
-// types so they stay aligned with what the underlying provider accepts.
-
-import type { ClaudeCodeOptions, CodexOptions } from "@ai-hero/sandcastle";
-
-export type ClaudeCodeSpec = {
-  provider: "claudeCode";
-  model: string;
-  promptPath: string;
-} & Pick<ClaudeCodeOptions, "effort">;
-
-export type CodexSpec = {
-  provider: "codex";
-  model: string;
-  promptPath: string;
-} & Pick<CodexOptions, "effort">;
-
-export type OpenCodeSpec = {
-  provider: "opencode";
-  model: string;
-  promptPath: string;
-};
-
-export type PiSpec = {
-  provider: "pi";
-  model: string;
-  promptPath: string;
-};
-
-export type AgentSpec = ClaudeCodeSpec | CodexSpec | OpenCodeSpec | PiSpec;
-
-// Each agent is keyed by purpose. Control flow (which agent runs when, the
-// docs-only branching, the conditional reviewer skip) stays in main.ts —
-// only the agent invocation parameters live here.
-//
-// Effort levels reflect Sandcastle's intended workload: simple bug fixes
-// and minor tweaks. Anything complex is handled HITL in Claude Code, not
-// here — so we don't pay for high-effort thinking on routine work.
-// Reviewer is the exception: edge-case stress-testing genuinely benefits
-// from extended thinking, and it's the last gate before a human review.
-export const AGENTS = {
-  implementer: {
-    provider: "claudeCode",
-    model: "claude-sonnet-4-6",
-    effort: "medium",
-    promptPath: "./.sandcastle/implement-prompt.md",
-  },
-  implementerDocs: {
-    provider: "claudeCode",
-    model: "claude-sonnet-4-6",
-    effort: "low",
-    promptPath: "./.sandcastle/implement-docs-prompt.md",
-  },
-  reviewer: {
-    provider: "claudeCode",
-    model: "claude-sonnet-4-6",
-    effort: "high",
-    promptPath: "./.sandcastle/review-prompt.md",
-  },
-  prOpener: {
-    provider: "claudeCode",
-    model: "claude-haiku-4-5-20251001",
-    effort: "low",
-    promptPath: "./.sandcastle/pr-prompt.md",
-  },
-} as const satisfies Record<string, AgentSpec>;
-
-export type AgentKey = keyof typeof AGENTS;
